requirements.txt
celery==5.4.0
Django==5.1
django-celery-beat==2.7.0
django-filter==24.3
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
pillow==10.4.0
psycopg2-binary==2.9.9
python==3.12.3
python-dotenv==1.0.1
django-stubs==5.0.4
coverage==7.6.1
django-cors-headers==4.4.0
redis==5.0.8
flake8==7.1.1
eventlet==0.36.1